Issue #98 - unexpected keycode
While investigating the issue, we discovered a keymap that contained
invalid keycodes. The keycodes happen to parse because they matched keys
but generated invalid keymaps.

We haven't figured out how to replicate this issue, short of editing the
JSON file itself.

As a defensive measure to prevent importing bad data, add another check
for the length of the keycode. We don't support keycodes shorter than 4
characters in length. In those cases set the key to KC_NO and warn
the user on the console.