Issue #98 - unexpected keycode in JSON
While investigating the issue, we discovered a keymap that contained
invalid keycodes. The keycodes happen to parse because they matched keys
but generated invalid keymaps.

This can be replicated by using an ANY key from keycodes and entering a
'.' or other valid keypress.js value. The UI would display a valid
keycode but the backend would receive a keypress.js value, causing the
compile to fail.

Fixes:

 - As a defensive measure, add another check for the length of the
   keycode. We don't support keycodes shorter than 4 characters in
   length. In those cases set the key to KC_NO and warn the user
   on the console.
 - prevent parsing code from parsing keys values.
   Adds a flag to prevent general non keypress.js uses from
   validating keypress.js aliases as valid keycodes