Init empty layers with KC_NO
We discussed changing the default to KC_TRNS but jackhumbert felt that
for configurator this was the wrong default.

If a key with a layer change is added to a keymap it only adds a single
key to that layer. This causes problems during export and causes the
compiler to fail.

 - new empty keys are force initialized to KC_NO for a new layer
 - any action that causes a layer to be created now forces
   initialization of that layer with all keys
 - in keymap class use initLayer function for consistency
 - Change layer behavior
   - when you change layer this initializes a layer
   - layer 0 is always non-empty
 - Hoist assignment out of loop