Upload info.json preview state bug (#212)
Separate the act of requesting a preview, from the state of being in a
preview. Input file doesn't send a cancel callback if the user opens a
file dialog but then cancels it. Add a new state where we preview is
requested which is only used to open the dialog and it is reset
immediately after.

Thanks to @noroadsleft for figuring this tricky state issue out.

Fix for issue #211