Fix keycount when clearing base layer
use the layout count rather than the keymap as we clear the base
layer and rely on the visual keymap to not send us incorrect
indices.  It may contain less than the entire keymap codes but
this is an internal detail that shouldn't be exposed to the UI.

The currently selected layout should always contain the correct number
of keys.

 - use a getter as this way it's cached until the layout or keyboard
 changes.

committed: #       modified:   src/components/Main.vue #       modified:
src/store/modules/app/getters.js # # Untracked files: #       CNAME