Use await to simplify callbacks in changeKeyboard
When I originally wrote this, await was still pretty new and how to use
it wasn't as clear to me as now. Flatten out the changeKeyboard function
so it takes advantage of await and remove the callbacks.

This should make the code a little bit more deterministic at the cost of
using await which uses more resources.