When switching between ISO and ANSI modify keymap
In the previous implementation the BaseKey rendering code was doing a
dynamic lookup to render the keycode for each key, so that we would
correctly render ISO keycodes.

This was fine until someone wanted to switch out the keycodes for locale
specific keycodes and then it required changes in many places in order
for the rendering to be correct for the keycodes.

Instead of doing it per key rendering, just sweep all the codes once
when we switch locale ISO/ANSI on master and update all the names to the
newly corrected names.

This will enable preconditions PR which allow support for all the
locales that QMK firmware supports in a more scaleable manner that
doesn't touch nearly as many places in the UI code.