fix: show the name instead of the code in print view
Issue [1363](https://github.com/qmk/qmk_configurator/issues/1363)

Change the PrintKey to show the `.name` instead of the `.code` attribute
because it's less human friendly. Remove some of the extraneous
formatting around the element.

This stems slightly from the different way we render Key components in
the printing vs editing.