Fix for `KC_NO` not triggering the InfoBar (#1435)
* Fix for `KC_NO` not triggering the InfoBar

PR #1428 introduced a check which allows hovering a keycode in the
Keycode Selection array to trigger the InfoBar only if the key being
hovered has a `name` property. This in turn caused hovering the `KC_NO`
keycode to no longer trigger the InfoBar, because its `name` property
was an empty string.

This commit sets the `name` property of the `KC_NO` keycode to 'N/A' in
order to bypass the aforementioned check, and thus trigger the InfoBar.

* Remove CSS rule for empty keycodes (per fauxpark)